requirements.txt
uvicorn[standard]==0.24.0
fastapi==0.110.0
python-dotenv==1.0.0
httpx==0.25.1
webrtcvad==2.0.10
openai==1.3.7
asyncpg==0.29.0
rapidfuzz==3.6.1
